+++ contracts/contracts/interfaces/IMagnifyCashV1.sol
@@ -46,7 +46,7 @@ interface IMagnifyCashV1 {
         uint256 amountPaidBack;
         address nftCollection;
         uint64 startTime;
-        uint64 nftId;
+        uint256 nftId;
         uint64 lendingDeskId;
         uint32 duration;
         uint32 interest;
@@ -181,7 +181,7 @@ interface IMagnifyCashV1 {
     function initializeNewLoan(
         uint64 _lendingDeskId,
         address _nftCollection,
-        uint64 _nftId,
+        uint256 _nftId,
         uint32 _duration,
         uint256 _amount,
         uint32 _maxInterestAllowed
